Added I/O Scheduler : Zen
ZEN is based on the Noop, Deadline and SIO I/O schedulers. It's an FCFS
(First come, first serve) based algorithm, but it's not strictly FIFO.
ZEN does not do any sorting. It uses deadlines for fairness, and treats
synchronous requests with priority over asynchronous ones. Other than
that, it's pretty much the same as Noop blended with VR features. 

Also-by: Slawek <yarpiin@gmail.com>